JSON Dynamic Data through hbs
I have modified the logic of the /travel page in the web app to iterate through the contents of a JSON file in order to serve a dynamic list of data to the user. This has been addressed up through the controller and is working in testing. I did add a few debugging statements to the shell so I could trace the logic during testing. These can be commented out before deployment.
-mdf